Criando uma RycleView

+++ app/src/main/java/com/example/aplicativoigor/model/Atividade.kt
@@ -5,7 +5,7 @@ import java.util.*
 data class Atividade (
     var descricao: String,
     var data: String,
-    var prioridade: Int,
+    var prioridade: Float,
     var tipoAtividade: String
 )
